feat: replaced old browser set with new one
