+++ src/app/sets/layout.tsx
@@ -1,5 +1,4 @@
 import Navbar from "@/components/layout/Navbar/Navbar";
-import { Card } from "@/components/ui";
 
 export default function SetsLayout({
   children,
@@ -7,10 +6,10 @@ export default function SetsLayout({
   children: React.ReactNode;
 }) {
   return (
-    <section className="max-h-screen">
+    <section>
       <Navbar displayLinks={false} />
-      <div className="mx-auto max-w-4xl px-4 pb-10">
-        <Card>{children}</Card>
+      <div className="mx-auto flex max-w-screen-2xl flex-col gap-6 px-4 md:flex-row">
+        {children}
       </div>
     </section>
   );
