+++ src/app/sets/loading.tsx
@@ -1,28 +1,37 @@
-import { Skeleton } from "@/components/ui";
+//@ts-nocheck
+
+import { Card, Divider, Skeleton } from "@/components/ui";
 
 const SetsLoading = async () => {
   return (
     <>
-      <div className="flex justify-between">
-        <Skeleton className="w-1/5" />
-        <Skeleton className="h-8 w-1/5" />
-      </div>
-      <div className="flex flex-col gap-3">
-        <Skeleton className="h-8 w-2/5" />
-        <Skeleton className="h-5 w-4/5" />
-      </div>
-      <div className="flex flex-row gap-4">
-        <Skeleton className="h-8 w-24" />
-        <Skeleton className="h-8 w-28" />
-        <Skeleton className="h-8 w-16" />
-      </div>
-      <div className="flex flex-col gap-3">
-        <Skeleton className="h-20 w-full" />
-        <Skeleton className="h-20 w-full" />
-        <Skeleton className="h-20 w-full" />
-        <Skeleton className="h-20 w-full" />
-        <Skeleton className="h-20 w-full" />
-      </div>
+      <aside className="static top-12 flex h-fit max-h-screen w-full flex-col gap-4 md:sticky md:w-full md:max-w-[18rem]">
+        <div className="flex items-center justify-between gap-12">
+          <Skeleton className="h-10 w-full" />
+
+          <Skeleton className="h-10 w-16" />
+        </div>
+        <Divider className="my-0" />
+        <div className="flex gap-3 md:flex-col">
+          <Skeleton className="h-10 w-full" />
+          <Skeleton className="h-10 w-full" />
+          <Skeleton className="h-10 w-full" />
+        </div>
+      </aside>
+      <Card className="w-full flex-col gap-4">
+        <div className="grid grid-flow-row-dense grid-cols-2 gap-4 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5">
+          <Skeleton className="aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="col-span-2 row-span-2 aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="col-span-2 row-span-2 aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="col-span-2 row-span-2 aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="col-span-2 row-span-2 aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="aspect-square h-full w-full rounded-lg" />
+          <Skeleton className="aspect-square h-full w-full rounded-lg" />
+        </div>
+      </Card>
     </>
   );
 };
