+++ src/app/sets/page.tsx
@@ -1,13 +1,29 @@
 /* eslint-disable @next/next/no-img-element */
-import SetSelector, {
-  SetSelectorProps,
-} from "@/features/sets/components/SetSelector";
+import { SetCategories, routes } from "@/constants";
+import SetSelector, { SetSelectorProps } from "@/features/xsets/SetSelector";
 import getUserDataServer from "@/lib/getUserDataServer";
 import { Database } from "@/types/supabase";
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
+import { redirect } from "next/navigation";
 
-const SetsPage = async () => {
+const SetsPage = async ({
+  searchParams,
+}: {
+  searchParams: { category: SetCategories; query?: string };
+}) => {
+  const { category, query } = searchParams;
+
+  await new Promise((resolve) => setTimeout(resolve, 5000));
+
+  if (
+    category !== "community" &&
+    category !== "featured" &&
+    category !== "personal"
+  ) {
+    redirect(routes.sets.browser("featured", query));
+  }
+
   const user = await getUserDataServer();
   const supabase = createServerComponentClient<Database>({ cookies });
